requirements.txt
python-dotenv==0.20.0
requests==2.27.1
pandas==1.3.4
tqdm==4.64.0
